+++ test/remote/gateways/remote_stripe_connect_test.rb
@@ -1,0 +1,42 @@
+require 'test_helper'
+
+class RemoteStripeConnectTest < Test::Unit::TestCase
+  def setup
+    @gateway = StripeGateway.new(fixtures(:stripe))
+
+    @amount = 100
+    @credit_card = credit_card('4242424242424242')
+    @declined_card = credit_card('4000000000000002')
+    @new_credit_card = credit_card('5105105105105100')
+
+    @options = {
+      :currency => "USD",
+      :description => 'ActiveMerchant Test Purchase',
+      :email => 'wow@example.com'
+    }
+  end
+
+  def test_application_fee_for_stripe_connect
+    assert response = @gateway.purchase(@amount, @credit_card, @options.merge(:application_fee => 12 ))
+    assert response.params['fee_details'], 'This test will only work if your gateway login is a Stripe Connect access_token.'
+    assert response.params['fee_details'].any? do |fee|
+      (fee['type'] == 'application_fee') && (fee['amount'] == 12)
+    end
+  end
+
+  def test_successful_refund_with_application_fee
+    assert response = @gateway.purchase(@amount, @credit_card, @options.merge(:application_fee => 12))
+    assert response.params['fee_details'], 'This test will only work if your gateway login is a Stripe Connect access_token.'
+    assert refund = @gateway.refund(@amount, response.authorization, :refund_application_fee => true)
+    assert_success refund
+    assert_equal 0, refund.params["fee"]
+  end
+
+  def test_refund_partial_application_fee
+    assert response = @gateway.purchase(@amount, @credit_card, @options.merge(:application_fee => 12))
+    assert response.params['fee_details'], 'This test will only work if your gateway login is a Stripe Connect access_token.'
+    assert refund = @gateway.refund(@amount - 20, response.authorization, { :refund_fee_amount => 10 })
+    assert_success refund
+  end
+
+end
